Rename model filenames to match class names
