+++ pylivoltek/models/__init__.py
@@ -20,16 +20,16 @@ from pylivoltek.models.current_power_flow import CurrentPowerFlow
 from pylivoltek.models.device import Device
 from pylivoltek.models.device_details import DeviceDetails
 from pylivoltek.models.device_list import DeviceList
-from pylivoltek.models.energy_storage_controller import StorageController
+from pylivoltek.models.storage_controller import StorageController
 from pylivoltek.models.energy_store import EnergyStore
 from pylivoltek.models.energy_store_battery_type import EnergyStoreBatteryType
 from pylivoltek.models.energy_store_history_map import EnergyStoreHistoryMap
 from pylivoltek.models.energy_store_history_map_item import EnergyStoreHistoryMapItem
 from pylivoltek.models.grid_import_export import GridImportExport
 from pylivoltek.models.grid_import_export_api_response import GridImportExportApiResponse
-from pylivoltek.models.grid_import_export_history import GridEnergyHistoryRecord
+from pylivoltek.models.grid_energy_history_record import GridEnergyHistoryRecord
 from pylivoltek.models.grid_import_export_list import GridImportExportList
-from pylivoltek.models.history_item import EnergyHistoryPoint
+from pylivoltek.models.energy_history_point import EnergyHistoryPoint
 from pylivoltek.models.inline_response200 import InlineResponse200
 from pylivoltek.models.inline_response2001 import InlineResponse2001
 from pylivoltek.models.inline_response2001_data import InlineResponse2001Data
@@ -37,14 +37,14 @@ from pylivoltek.models.inline_response2002 import InlineResponse2002
 from pylivoltek.models.inline_response2003 import InlineResponse2003
 from pylivoltek.models.inline_response2004 import InlineResponse2004
 from pylivoltek.models.inline_response2005 import InlineResponse2005
-from pylivoltek.models.inline_response2005_data import DeviceApiResponse
+from pylivoltek.models.device_api_response import DeviceApiResponse
 from pylivoltek.models.inline_response2006 import InlineResponse2006
 from pylivoltek.models.inline_response2007 import InlineResponse2007
-from pylivoltek.models.inline_response2008 import RecentUtilityEnergyResponse
-from pylivoltek.models.inline_response2008_data import UtilityEnergyTotals
-from pylivoltek.models.inline_response2008_data_etotal_to_grid import EnergyMeasurement
-from pylivoltek.models.inline_response2009 import RecentSolarEnergyResponse
-from pylivoltek.models.recent_grid_import_export import SiteEnergyReading
+from pylivoltek.models.recent_utility_energy_response import RecentUtilityEnergyResponse
+from pylivoltek.models.utility_energy_totals import UtilityEnergyTotals
+from pylivoltek.models.energy_measurement import EnergyMeasurement
+from pylivoltek.models.recent_solar_energy_response import RecentSolarEnergyResponse
+from pylivoltek.models.site_energy_reading import SiteEnergyReading
 from pylivoltek.models.site import Site
 from pylivoltek.models.site_list import SiteList
 from pylivoltek.models.site_overview import SiteOverview
