+++ pylivoltek/models/grid_energy_history_record.py
